requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0

python-decouple==3.8

google-auth==2.43
requests==2.32.3
gunicorn==23.0.0
boto3==1.38.15